# Avoid Index DeprecationWarning in Series getitem