--- a/pandas/core/internals/managers.py
+++ b/pandas/core/internals/managers.py
@@ -1505,7 +1505,11 @@ class SingleBlockManager(BlockManager):
         if axis >= self.ndim:
             raise IndexError("Requested axis not found in manager")
 
-        return type(self)(self._block._slice(slobj), self.index[slobj], fastpath=True)
+        return type(self)(
+            self._block._slice(slobj),
+            self.index._getitem_deprecate_nd(slobj, warn=False),
+            fastpath=True,
+        )
 
     @property
     def index(self):