--- a/pandas/core/internals/managers.py
+++ b/pandas/core/internals/managers.py
@@ -4,6 +4,7 @@ import itertools
 import operator
 import re
 from typing import Dict, List, Optional, Sequence, Tuple, Union
+import warnings
 
 import numpy as np
 
@@ -1505,11 +1506,13 @@ class SingleBlockManager(BlockManager):
         if axis >= self.ndim:
             raise IndexError("Requested axis not found in manager")
 
-        return type(self)(
-            self._block._slice(slobj),
-            self.index._getitem_deprecate_nd(slobj, warn=False),
-            fastpath=True,
-        )
+        with warnings.catch_warnings():
+            warnings.filterwarnings(
+                "ignore", "Support for multi-dim", DeprecationWarning
+            )
+            index = self.index[slobj]
+
+        return type(self)(self._block._slice(slobj), index, fastpath=True,)
 
     @property
     def index(self):