@TomAugspurger 2020-01-27
Oh, I got who calls what backwards here... Will need to re-think things...

We don't want to change all the index subclasses `__getitem__`s to do the `_getitem_deprecate_nd` version that maybe silences warnings.