@TomAugspurger 2020-01-27
Changed to using a warnings filter. Note that this has a perf impact for slicing a Series. ~25% slower

Master

```
In [9]: s = pd.Series(range(10000))

In [10]: %timeit s[:5]
43.5 µs ± 2.15 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [11]: t = pd.Series([], dtype=float)

In [12]: %timeit t[:]
47.8 µs ± 2.24 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

This PR

```
In [7]: s = pd.Series(range(10000))

In [8]: %timeit s[:5]
55.6 µs ± 1.69 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [9]: t = pd.Series([], dtype=float)

In [10]: %timeit t[:]
64.2 µs ± 2.27 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

Is that slowdown acceptable? Right now, I don't think it is, given that

1. This is just a DeprecationWarning
2. We'll be deprecating this in the future anyway.