@jbrockmendel 2020-01-27
I think there are 2 ways we get to get_slice with non-slice.  One of them is L931, and can be avoided be extracting the single-item list after potentially warning in #31333.  The other is on L940 and is explicltly  an "mpl hackaround".  Instead of calling self._get_values there we could do something like `return self._values[key]`