@TomAugspurger 2020-01-28
Thanks @jbrockmendel, that did it. We're able to apply the filter only when using the mpl compat path.

master

```python
In [2]: s = pd.Series(range(10000))

In [3]: %timeit s[:, None]
36.2 µs ± 293 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [4]: t = pd.Series([], dtype=float)

In [5]: %timeit t[:]
52 µs ± 3.5 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [6]: %timeit s[:5]
45 µs ± 2.27 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
``` 



PR

```python
In [4]: s = pd.Series(range(10000))

In [5]: %timeit s[:, None]
50.2 µs ± 2.56 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [6]: t = pd.Series([], dtype=float)

In [7]: %timeit t[:]
48.4 µs ± 2 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [8]: %timeit s[:5]
44.3 µs ± 2.52 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```