@TomAugspurger 2020-01-28
Had to xfail a test asserting that this produced a warning. Will want to restore that with one from `Series.__getitem__` I think, but not for 1.0