@jbrockmendel 2020-01-28
> So the first case is the only one with a slowdown, which I think is acceptable given that we want to deprecate that behavior anyway.

Sounds good